Update Smart Blind Stick
Blind peoples have to face many challenges in their life, one of them is finding their way on the streets. On the streets, there are so many vehicles and obstacles that may block their way and also may injure them.

So keeping this problem in the mind we developed a Smart blind stick that scans for the obstacles in front of it with the help of an ultrasonic sensor.